+++ Ampel/Ampel.vcxproj
@@ -138,6 +138,7 @@
     <ClInclude Include="LightBody.h" />
     <ClInclude Include="LightBox.h" />
     <ClInclude Include="RoadLights.h" />
+    <ClInclude Include="RoadsDB.h" />
     <ClInclude Include="Setup.ipp" />
     <ClInclude Include="TrafficLight.h" />
     <ClInclude Include="VehicleLight.h" />
