+++ Ampel/Application.h
@@ -1,20 +1,7 @@
 #pragma once
 
 #include "Includes.h"
-#include "RoadLights.h"
-
-
-struct Road
-{
-	Road(sdl::Renderer* r, const sdl::Point<int>& pos)
-		: light(r, pos)
-	{
-	}
-
-	RoadLights light;
-	//std::chrono::steady_clock::time_point onTill;
-	//std::vector<size_t> path;
-};
+#include "RoadsDB.h"
 
 
 class App
@@ -23,19 +10,6 @@ class App
 
 public:
 	App(sdl::Renderer* r);
-
-	Road* insideWhich(const sdl::Point<int>& p) noexcept
-	{
-		Road* ptr = nullptr;
-		for (auto iter = m_lights.rbegin(); iter != m_lights.rend(); ++iter)
-			if (sdl::collision(iter->light.shape(), p))
-			{
-				ptr = &*iter;
-				break;
-			}
-
-		return ptr;
-	}
 
 	void input(const SDL_Event& e)
 	{
@@ -56,7 +30,7 @@ private:
 	sdl::Renderer* m_r;
 
 	std::unique_ptr<sdl::IState> m_state;
-	std::vector<Road> m_lights;
+	RoadsDB m_roads;
 };
